feat: make logger type configurable
